test: Fix gSSP in base test rig
