+++ packages/e2e/src/pages/pages/useQueryState/dynamic/[route]/index.tsx
@@ -1,16 +1,3 @@
-import { GetServerSideProps } from 'next'
-import { parseAsString } from 'next-usequerystate'
-import IntegrationPage from '../../index'
+import IntegrationPage, { getServerSideProps } from '../../index'
 export default IntegrationPage
-
-export const getServerSideProps = (async ctx => {
-  const string = parseAsString.parseServerSide(ctx.query.string)
-  console.dir({ string })
-  return {
-    props: {
-      string
-    }
-  }
-}) satisfies GetServerSideProps<{
-  string: string | null
-}>
+export { getServerSideProps }
